The free-server queue tests now all pass.
Changed the free-server queue from an atom to a ref because I needed
couldn't figure out how to extract the next item from the queue and then
remove that item, all in an atomic fashion with swap! I also had to put
a doall in front of the pmap that inserts all the servers into the
free-server queue; otherwise the tests continued when we didn't actually
have all the parallel threads finished and the servers in the queue yet.